Add helpers for creating transports (#2029)
Motivation:

Creating an HTTP/2 transport requires quite a bit of typing and knowing
what types to use. We can make use of type inference to make this a bit
easier for users.

Modifications:

- Add constrained extensions to `ClientTransport` and `ServerTransport`
allowing for `http2NIO` and `http2NIOTS` transports to be created via
static methods. This makes it possible for transports to be created
inline in the `init` of a `GRPCClient` or `GRPCServer` without the users
need the full type.
- Extend the `defaults` functions so that they can take an optional
trailing closure to further modify the defaults. This allows for configs
to be created and modified inline when making a transport.
- A config is now required to create client transports and the NIOTS
server transport as these will soon require a transport security config.

Results:

Easier to create clients/servers