Update plugin bundling script
Motivation:

The plugin bundling script which is used when we do a release doesn't
update the dependencies so will just use whatever version is currently
resolved. This means protobuf might not be up-to-date and an older
version of `protoc-gen-swift` might get build.

Modifications:

- Run package update before building

Result:

The bundled `protoc-gen-swift` should be up-to-date