Add server stream events
Motivation:

As a service author it's useful to know if the underlying stream has
been cancelled (because it's timed out, the remote peer closed it, the
connection dropped etc).

For cases where the stream has already closed this can be surfaced by a
read or write failing. However, for cases like server-streaming RPCs
where there are no reads and writes can be infrequent it's useful to
have a more explicit signal.

Modifications:

- Add a `ServerStreamState` which provides two API for this:
  - An explicit "has this RPC been cancelled?"
  - An async sequence of events which can happen the underlying stream
- Add the stream state to the server context
- Update the in-process transport to use this

Result:

Server context includes a way for users to know if their RPC has been
cancelled.