Restructure in-process transport (#2071)
Motivation:

To create a connected pair of `ServerTransport` and `ClientTransport`,
users need to call `InProcessTransport.makePair(serviceConfig:)` which
returns a tuple. The spelling of this API can be more intuitive and
simple.

Modifications:

- Restructure `InProcessTransport` to include two properties: `server`
(the `ServerTransport`) and `client` (the `ClientTransport`), and an
`init(serviceConfig:)` that pairs these properties.
- Add missing GRPCInProcessTransport dependency to GRPCCore test target.

Result:

The spelling of the `InProcessTransport` API will be better.