Add an RPC cancellation handler (#2090)
Motivation:

As a service author it's useful to know if the RPC has been cancelled
(because it's timed out, the remote peer closed it, the connection
dropped etc).

For cases where the stream has already closed this can be surfaced by a
read or write failing. However, for cases like server-streaming RPCs
where there are no reads and writes can be infrequent it's useful to
have a more explicit signal.

Modifications:

- Add a `ServerCancellationManager`, this is internal per-stream storage
for registering cancellation handlers and storing whether the RPC has
been cancelled.
- Add the `RPCCancellationHandle` nested within the `ServerContext`.
This holds an instance of the manager and provides higher level APIs
allowing users to check if the RPC has been cancellation and to wait
until the RPC has been cancelled.
- Add a top-level `withRPCCancellationHandler` which registers a
callback with the manager.
- Add a top-level `withServerContextRPCCancellationHandle` for creating
and binding the task local manager. This is intended for use by
transport implementations rather than users.
- Update the in-process transport to cancel RPCs when shutting down
gracefully.
- Update the server executor to cancel RPCs when the timeout fires.

Result:

Users can watch for cancellation using `withRPCCancellationHandler`.