Avoid using vendored lock in BroadcastAsyncSequence (#2108)
Motivation:

ManagedBuffer is explicitly not 'Sendable' on Swift nightly builds. As a
result, our Lock type stops being 'Sendable'. This is only used by the
'BroadcastAsyncSequence'.

Modifications:

- Modify 'BroadcastAsyncSequence' to use 'Mutex'. One downside to this
approach is that when a subscribe wants to consume an element and
there's no element available, the lock must be released and then
re-acquired once a continuation has been created.
- Remove our lock, and update NOTICES.txt to reflect the fact we no
longer use a copy of NIOs lock.

Result:

- Less code
- CI passes