Add RPCErrorConvertible
Motivation:

If an error is thrown from a server RPC then the status sent to the
client will always have the unknown error code unless an `RPCError` is
thrown.

Moreover, there are various extensions to gRPC which rely on additional
information being stuffed into the metadata. This is difficult and a bit
error prone for users to do directly.

We should provide a mechanism whereby errors can be converted to an
`RPCError` such that the appropriate code, message and metadata are sent
to the client.

Modifications:

- Add the `RPCErrorConvertible` protocol. Conforming types provide
  appropriate properties to populate an `RPCError`.
- Add handling for this in the server executor such that convertible
  errors are converted into an `RPCError`.

Result:

Easier for users to propagate an appropriate status