Simplify code gen lib interface (#2169)
Motivation:

When reviewing APIs I noticed a few inconsistencies and some
abstractions which are more complicated than they need to be. We should
simplify this before we commit to a stable API.

Modifications:

- Replace the 'Name' type with a more concrete 'ServiceName' and
'MethodName'. They carry approximately equivalent information but the
information is expressed in terms of what the names are used for (i.e. a
function vs. what the expected format of the name is, e.g. lowercase).
This makes it easier for users to understand how the names will be used
and leaves room for more customisation in the future.
- The service descriptor used two names: a namespace name and a service
name. These have now been compressed into a single object (not all
namespace name values were used). The namespace name was never used in
isolation so this was adding unnecessary complexity.
- The top-level 'SourceGenerator' type has been renamed 'CodeGenerator'.
This is consistent with being a 'CodeGen' module and having a
'CodeGenerationRequest'.
- The closures for returning code snippets to create a
serializer/deserializer were renamed to make their user clearer.
- Accidental public API was removed.

Result:

Slightly smaller and more consistent API.