Add a new field, transportSpecific to ServerContext
Motivation:

Currently there's no way to plumb through details from the transport
level to a request handler. Adding this field allows transports, such as
the nio transport, to add the peer certificate to the server context
when using mTLS.

From there there it's easy for an interceptor to take this data and
propogate it forward to a request handler.

Modifications:

This PR adds a single field to the ServerContext that transports can use

Result:

A new field will be accessible to transports and consumers of the API