Add a dev-tool subpackage (#2167)
Motivation:

A number of test in this package and others rely on ad-hoc services
using Codable. This is less overhead than using protobuf as you it's not
always available. It also means the messages are defined in Swift so
they're easy to change without needing to regenerate. However, the
service glue code is hand rolled. We can avoid this by having a little
adapter sit on top of the code gen lib.

Modifications:

- Add a grpc-dev-tool package to dev. We can use this as a place to add
tooling and other helpers without worrying about worsening the
experience for end users (because of additional dependencies, more
public API and so on).
- For now this has a single executable for generating code from a JSON
config file. The schema for the services is limited, but that's fine,
it's not a general purpose tool.

Result:

- We have a tool which can generate grpc code from a JSON definition
which uses Codable message types.