Improve error message when client is run when shutdown (#2215)
Motivation:

GRPCClient can throw an error when run is called more than once or if
run is called after it has been shutdown. Normally this would happen if
a user caller 'run()' more than once, but can also happen if
'withGRPCClient' is called and the client is never used and the body
returns quickly.

Modifications:

- Improve the error message

Result:

Better error message.