fix generated method names (#2224)
This PR fixes incorrectly generated method names.

For example:
`rpc GRPCUnary(google.protobuf.Empty) returns (google.protobuf.Empty)
{}`

The plugin correctly generates:
- `makeGRPCUnaryInterceptors`
- `Methods.gRPCUnary`
- `func gRPCUnary`

but incorrectly:
- `makeGrpcunaryCall`
This seems to be the only place with a different logic.

And when using lowercase letters, the interceptor method name is
incorrect:
`rpc unary(google.protobuf.Empty) returns (FunctionName) {}`

The plugin correctly generates:
- `makeUnaryCall`
- `Methods.unary`
- `func unary`
but incorrectly:
- `makeunaryInterceptors`

The PR fixes these two issues by introducing a common logic for
`methodMakeFunctionCallName` and `methodInterceptorFactoryName`. The
logic is similar to the one in `methodFunctionName`, just that it
uppercases the first character instead of lowercasing.

Thus, for the given examples the result is:
`makeGRPCUnaryCall` and `makeUnaryInterceptors`

---------

Co-authored-by: George Barnett <gbarnett@apple.com>